pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "spengine"
version = "1.0.0"
description = "Mini Input Output Tool"
authors = [
    { name="Alif", email="yoyojr404@gmail.com" }
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = ["kafka-python", "boto3", "GeoAlchemy2", "greenstalk", "jmespath", "loguru", "openpyxl", "pandas", "pika", "Pillow", "psycopg2-binary", "pydantic", "pymongo", "python-dotenv", "python-ssdb", "requests", "SQLAlchemy", "srsly", "elasticsearch"]

[project.scripts]
spengine = "spengine.main:main"
